Revert "Revert "[front] Allow multiple wrappers to run and fetch resources (#9682)" (#9738)"
This reverts commit 9fc6c81119a4e587094756d6d247593bd892390a.

+++ front/pages/api/w/[wId]/spaces/[spaceId]/data_sources/[dsId]/tables/[tableId]/index.ts
@@ -32,7 +32,7 @@ async function handler(
   req: NextApiRequest,
   res: NextApiResponse<WithAPIErrorResponse<PatchTableResponseBody>>,
   auth: Authenticator,
-  space: SpaceResource
+  { space }: { space: SpaceResource }
 ): Promise<void> {
   const { tableId, dsId } = req.query;
 
@@ -123,6 +123,16 @@ async function handler(
       return;
 
     case "DELETE":
+      if (!dataSource.canWrite(auth)) {
+        return apiError(req, res, {
+          status_code: 403,
+          api_error: {
+            type: "data_source_auth_error",
+            message: "You are not allowed to update data in this data source.",
+          },
+        });
+      }
+
       const delRes = await deleteTable({
         owner: auth.getNonNullableWorkspace(),
         dataSource,
@@ -168,5 +178,5 @@ async function handler(
 }
 
 export default withSessionAuthenticationForWorkspace(
-  withResourceFetchingFromRoute(handler, "space")
+  withResourceFetchingFromRoute(handler, { space: { requireCanRead: true } })
 );
